separate wrappers for every module
